pulling sql connector files out to make each file more focused
